+++ README.md
@@ -1,2 +1,10 @@
 [![Build Status](https://quakecoresoft.canterbury.ac.nz/jenkins/job/qcore/badge/icon?build=last:${params.ghprbActualCommit=master)](https://quakecoresoft.canterbury.ac.nz/jenkins/job/qcore)
 
+## Installation:
+
+To install run `pip install -e .` from the root directory of this repository
+
+### Downloading data after install
+
+After installing, if you intend to use GMT for plotting you will need to run "download_data.py" to download and
+unpack the data used for these scripts
